Revise Build Instructions
I added the link to the video @cellos51 made about compiling the project. I updated the build instructions to reflect the steps made in the video and condensed the information to be more readable. I also tried fixing the dynamic badges, but I don't think it completely worked. They update automatically but it takes a really long time (couple hours) because of GitHub API Caching Issues. This is a small pull request, I'll open requests in the future for larger scale commits. @MeirGavish can merge if they want if cellos51 is busy. Let me know if I made an error, thanks! 👍